move src/math.ts to src/lib/
